feat(#62):객관식 연결(디자인 반영 x)
