+++ src/services/api.ts
@@ -158,34 +158,66 @@ export const login = async (request: LoginRequest): Promise<LoginResponse> => {
   }
 };
 
-export const createQuestions = async (
-  plainText: string
-): Promise<Question[]> => {
+export const createQuestions = async (text: string, type?: "TRUE_FALSE" | "MULTIPLE_CHOICE"): Promise<Question[]> => {
   try {
-    const token = getAccessToken();
+    const requestBody = {
+      plainText: text,
+      type: type || "TRUE_FALSE",
+    };
+    
+    console.log("API 요청 URL:", `${API_BASE_URL}/v1/clova/question`);
+    console.log("API 요청 본문:", requestBody);
+
+    // 기본 헤더만 설정
     const headers: Record<string, string> = {
       "Content-Type": "application/json",
     };
 
-    if (token) {
-      headers["Authorization"] = `Bearer ${token}`;
+    // 로그인된 사용자만 Authorization 헤더 추가
+    const token = getAccessToken();
+    if (token && token.trim() !== "") {
+      headers.Authorization = `Bearer ${token}`;
+      console.log("Authorization 헤더 포함됨:", `Bearer ${token}`);
+    } else {
+      console.log("비회원 요청 - Authorization 헤더 없음");
     }
+
+    console.log("최종 API 요청 헤더:", headers);
 
     const response = await fetch(`${API_BASE_URL}/v1/clova/question`, {
       method: "POST",
       headers,
-      body: JSON.stringify({
-        plainText,
-        type: "TRUE_FALSE",
-      } as CreateQuestionRequest),
+      body: JSON.stringify(requestBody),
     });
 
+    console.log("API 응답 상태:", response.status);
+    console.log("API 응답 헤더:", Object.fromEntries(response.headers.entries()));
+
     if (!response.ok) {
-      throw new Error(`HTTP error! status: ${response.status}`);
+      let errorText = "";
+      try {
+        errorText = await response.text();
+        console.error("API 에러 응답 본문:", errorText);
+      } catch (e) {
+        console.error("API 에러 응답 본문 읽기 실패:", e);
+      }
+      
+      throw new Error(`HTTP error! status: ${response.status}, message: ${errorText}`);
     }
 
     const data = await response.json();
-    return data;
+    console.log("API 성공 응답:", data);
+    
+    let questionData: any = data;
+    if (data && typeof data === 'object' && !Array.isArray(data)) {
+      if (data.questions && Array.isArray(data.questions)) {
+        questionData = data.questions;
+      } else if (data.data && Array.isArray(data.data)) {
+        questionData = data.data;
+      }
+    }
+    
+    return questionData || data;
   } catch (error) {
     console.error("Error creating questions:", error);
     throw error;
